change interface for event handling
Previously SDL_PollEvent used a _ptr o object to return the event in
every call, but this means allocating a SDL_Event structure in every
call; considering SDL_PollEvent is usually called once for every frame,
this can result in poorer performance and more GC pressure. Changed the
interface to use a pointer to a pre-allocated SDL_Event, and added a
next-event function that allocates as before, for simpler programs.